Convert image width and height (used by tiling) to float once.
The storage cost is the same, so might as well do this when building the pipeline instead of when running it.  This also avoids the awkward cvtsi2ss instruction that screws with register renaming.

CQ_INCLUDE_TRYBOTS=skia.primary:Test-Ubuntu-GCC-GCE-CPU-AVX2-x86_64-Release-SKNX_NO_SIMD

+++ src/opts/SkRasterPipeline_opts.h
@@ -796,12 +796,12 @@ SI SkNf mirror(const SkNf& v, float l/*imit*/) {
     result = SkNf::Min(result, nextafterf(l, 0));
     return assert_in_tile(result, l);
 }
-STAGE( clamp_x) { r = clamp (r, *(const int*)ctx); }
-STAGE(repeat_x) { r = repeat(r, *(const int*)ctx); }
-STAGE(mirror_x) { r = mirror(r, *(const int*)ctx); }
-STAGE( clamp_y) { g = clamp (g, *(const int*)ctx); }
-STAGE(repeat_y) { g = repeat(g, *(const int*)ctx); }
-STAGE(mirror_y) { g = mirror(g, *(const int*)ctx); }
+STAGE( clamp_x) { r = clamp (r, *(const float*)ctx); }
+STAGE(repeat_x) { r = repeat(r, *(const float*)ctx); }
+STAGE(mirror_x) { r = mirror(r, *(const float*)ctx); }
+STAGE( clamp_y) { g = clamp (g, *(const float*)ctx); }
+STAGE(repeat_y) { g = repeat(g, *(const float*)ctx); }
+STAGE(mirror_y) { g = mirror(g, *(const float*)ctx); }
 
 STAGE(save_xy) {
     auto sc = (SkImageShaderContext*)ctx;
